requirement.txt
numpy==2.2.2
opencv-python==4.11.0.86
pandas==2.2.3
pdf2image==1.17.0
Pillow==11.1.0
pytesseract==0.3.13
streamlit==1.41.1
streamlit-cropper==0.2.2
